+++ dev/src/ApiComparator/MethodFindingTraverser.php
@@ -1,0 +1,38 @@
+<?php
+/**
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+namespace Google\Cloud\Dev\ApiComparator;
+
+use PhpParser\Node;
+use PhpParser\Node\Stmt\ClassMethod;
+use PhpParser\NodeTraverser;
+use PhpParser\NodeVisitorAbstract;
+
+/**
+ * Traverse an AST and find class methods.
+ */
+class MethodFindingTraverser extends NodeVisitorAbstract
+{
+    public $methods = [];
+
+    function enterNode(Node $node) {
+        if ($node instanceof ClassMethod) {
+            $this->methods[] = $node;
+            return NodeTraverser::DONT_TRAVERSE_CHILDREN;
+        }
+    }
+}
